pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ikon-auth"
version = "0.1.4"
description = "A reusable OAuth2 token provider and JWT verification dependency for FastAPI microservices with JWKS support."
readme = "README.md"
authors = [
  { name = "Subhadip Hazra", email = "subhadip.hazra@keross.com" }
]

license = "MIT"
requires-python = ">=3.10"

keywords = [
    "OAuth2",
    "JWT",
    "FastAPI",
    "JWKS",
    "Authentication",
    "Security",
    "Authorization",
    "Microservices",
    "Access Token",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Framework :: FastAPI",
    "Operating System :: OS Independent",
]

dependencies = [
  "requests>=2.0.0",
  "python-jose[cryptography]>=3.3.0",
  "cryptography>=41.0.0",
  "pydantic-settings>=2.0.0",
  "pydantic>=2.0.0",
  "fastapi>=0.110",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/your-org/ikon-auth"
Repository = "https://github.com/your-org/ikon-auth"
Issues = "https://github.com/your-org/ikon-auth/issues"
Documentation = "https://github.com/your-org/ikon-auth#readme"
